Bugfix: Don't mutate fragments that are attached to query documents
This fixes an issue that arises when you perform the same query multiple times, and that query has fragments.

Because `addFragmentsToDocument` _appends_ fragments to `queryDoc.definitions`, each time the query is performed, the query document gains a new copy of each fragment.  This is particularly egregious, since `graphql-tag` maintains an identity map for queries