[docs] Use package-based Hexo theme, rather than a Git submodule. (#3148)
* docs: Make `scripts` consistent across "docs" repos.

For example, the `test` and `start` commands.

* docs: Use `devDependencies` instead of `dependencies`.

It's safe to mark all dependencies of this theme as 'dev' dependencies, as
none of them are needed to be installed in a production setting.

This was done automatically, so in some cases, this merges existing
`devDependencies`, and may also have added a missing `nodemon` `devDependency`
in cases where it was missing.  In other cases, this commit may not have
made such a change.

* docs: Fix top-level `hexo` versions in docs `package.json`s.

This is just metadata used by Hexo, but it's meant to be updated and committed.

* docs: Remove _config.yml options now covered by global theme config.

* docs: Add `_multiconfig.yml`, a build artifact, to docs `.gitignore`.

* docs: Use theme from npm, rather than submodule, using `chexo`.

The theme which the docs in this repository are based on
([`hexo-theme-meteor`](https://github.com/meteor/hexo-theme-meteor)) has been
published to npm.

Additionally, most of the configuration which was once present in this
repository's `_config.yml` file, has been moved into an Apollo-centric theme
configuration npm,
[`apollo-hexo-config`](https://www.npmjs.com/package/apollo-hexo-config) (and
similarly, [GitHub](https://github.com/apollographql/apollo-hexo-config)).

The theme bundled in this npm is a converged version of the two `apollo`
and `master` branches on the `hexo-theme-meteor` repository, which is a result
of the work undertaken in https://github.com/meteor/hexo-theme-meteor/pull/51.

Lastly, this makes use of a new `chexo` npm which acts as a wrapper for `hexo`
but permits the use of an inherited configuration from npm, rather than a
statically defined CLI file parameter (previously with `--config a,b`).

* docs: Upgrade Hexo to 3.6.0.

* docs: Upgrade Hexo generator/tag/server dependencies.

* docs: Add Renovate configuration.

While some repositories which host Hexo docs alongside other code may not have
Renovate setup, it's anticipated that they will in the future, at the very
least for the docs portions of the respective repositories.

By having this `renovate` section in the `package.json` for the Hexo docs, we
can ensure that they will receive the proper treatment and minimize the amount
of manual work necessary on those repositories.

Additionally, this ensures that there is a single point of upstream Apollo-docs
Renovate configuration through the use of the
[`renovate-config-apollo-docs`](https://npm.im/renovate-config-apollo-docs)
package, which inherits further from
[`renovate-config-mdg-docs`](https://npm.im/renovate-config-mdg-docs).

Repositories which already have Renovate setup should benefit automatically
from this new docs-specific configuration.