Update issue triage doc to better align with repo triage flow
The initial triage doc was copied from the Meteor repo, which
follows a different triage workflow than this repo does.
This commit updates the triage doc so that it better represents
this repo's current triage workflow. The filename has also been
changed to upper snake case to better align with the other repo
meta docs.