Remove `rxjs-compat` use
While merging in https://github.com/apollographql/apollo-client/pull/3500
we encountered `rxjs` `from` issues in some of the tests. I added
`rxjs-compat` to work around these issues, but this shouldn't be
needed. This commit aims to revert the `rxjs-compat` changes, to
see if the CI gods will now be appeased.