Bump graphql version to 14.0.0
So graphql are transitioning to real semver finally. This should fix #3562

I've set the devDependency to graphql@14.0.0-rc.2 to be able to be able
to run the tests on the new version and make sure everything's fine. But
I'm not sure if you think its ok to merge with that release candidate.

Also I think the semver in peerDependencies is more verbose than it
should be:

```js
"peerDependencies": {
    "graphql": "^0.11.0 || ^14.0.0"
}
```
This should cover all of the graphql versions from 0.11 and up, since
["^" includes minor and patch
versions](https://bytearcher.com/articles/semver-explained-why-theres-a-caret-in-my-package-json/)