Remove outdated README section
The Apollo Boost README config options section, and provided
config example, are outdated and have caused some confustion for
people. Since we already have everything up to date in the
official docs, this commit removes the duplicated (and invalid)
section from the README, replacing it with a link to the docs.

Fixes #3210.