Avoid WeakMap for InMemoryCache#transformDocumentCache.
Just in case the query DocmentNode is frozen for some reason, which would
prevent WeakMap polyfills from assigning it a unique ID property.

Query documents are relatively long-lived objects, and not extremely
numerous compared to other kinds of objects, so the risk of memory leaks
is minimal here, and the lifetime of this cache is the same as that of the
InMemoryCache object.