Stop using cache results in local resolver root values (#4495)
In an effort to simulate `defaults` behaviour from
`apollo-link-state`, we're leveraging cache results when
running local resolvers. The idea being that if a local resolver
isn't defined for a `@client` field, the local resolver
handling code can then fallback on using any matching cache
results, to resolve the field. While this works in theory,
it has introduced a few problems, like the one reported in #4474.

Since local resolvers adhere to Apollo Client's query fetch
policy, by defalut the cache is consulted first, when trying
to resolve a `@client` field. This means we shouldn't need to
attempt to resolve from the cache again, when processing local
resolvers, in most cases. There are a few situations where we
might want to do this, but the requirements are theoretical at
this point, and can be addressed in future changes if needed.

This commit removes the extra cache check, and adds
a test to verify that the behaviour reported in #4474 is fixed.