Avoid handling client typeDefs in LocalState class.
The only reason we currently store/manage client typeDefs in the Apollo
Client codebase is so that they can be consumed by the Apollo DevTools in
development.

Because the DevTools do not have a public API, it's important for
developers to continue passing any client typeDefs to the ApolloClient
constructor, but we do not have to expose those typeDefs through any
convenient public API, and the LocalState class should not need to know
about them at all, much less perform any expensive normalization, since
all of that work can be done by the DevTools.

Instead, the typeDefs are now exposed as client.typeDefs, exactly as they
were originally passed to the ApolloClient constructor. Local resolvers
will no longer receive context.schemas, but that's no great loss because
the LocalState implementation is new in apollo-client@2.5.0. It would be
much harder to remove that functionality after shipping v2.5.0, which is
why it's important to do it now.

Thanks to this change, we're back to having zero imports of
graphql/language/printer in the Apollo Client codebase (though the printer
is still used by apollo-link).

@hwillson @justinanastos I realize there's been some churn in the way the
DevTools and the client communicate recently, but the good news is we can
iterate freely because it's a private API. Let me know if you have any
questions about this change!