Apply result caching to StoreReader#execSubSelectedArray, too.
Should help with #4662.

The caching is currently keyed on the identity of the input array object,
rather than the identities of the elements. It would be relatively easy to
use cacheKeyRoot.lookupArray(array) instead of just array in makeCacheKey,
which would enable preserving array identity in even more cases, but that
approach would also make the cache key computation more expensive, since
makeCacheKey would have to traverse an input array of unbounded size.