Bring in React Apollo source
This commit kicks of the initial merge of React Apollo into
Apollo Client. It merges in the source of:

- @apollo/react-common
- @apollo/react-hooks
- @apollo/react-testing

The HOC (@apollo/react-hoc) and components (@apollo/react-components)
packages are not going to be brought into Apollo Client, and
the React SSR package (@apollo/react-ssr) will also be kept separate
for now.

This commit represents the initial merge of the code (with changes
required to get all tests passing), but does not represent the
final state of Apollo Client's React integration. Now that the
RA code is part of AC, work on leveraging the tighter RA/AC
integration can proceed.