Re-arrange core link code to better reflect the implementation
I'm not sure what the historical reason was for exposing link
utility functions like `concat`, `empty`, `from`, `split` and
`execute` both as static functions on `ApolloLink` and as
separate importable functions. Since it might be a bit jarring
for people if we remove one of these options, for now lets
move all `ApolloLink` functionality back into the `ApolloLink`
class, and instead re-export that functionality as separate
utility functions. Using the utility functions through
`ApolloLink` seems to be the most popular approach, and by
aligning this in the code we'll make future maintenance a bit
easier.