Unify Field{Read,Merge}Options as much as possible, for consistency.
Although it was tempting to make the read function feel like a GraphQL
resolver function (parentObject, args, context, info), that signature
would not have been appropriate for the merge function, and having to
remember two totally different signatures for reading and merging did not
seem ideal.

Now, the only difference between the signatures is that the merge function
takes incoming data, to be merged with the existing data (if any):

  read(existing, { args, parentObject, field, variables })
  merge(existing, incoming, { args, parentObject, field, variables })

It would have been nice to use named options for everything, including the
existing and incoming parameters, but it's important for those parameters
to be positional so that the developer can specify their types, without
also having to provide a new type for the entire options object.