Make mocking utilities require a Promise reject function.
Because Jest has no ability to trace uncaught exceptions back to the test
responsible for throwing them, the Apollo Client test suite has been
plagued for a long time by "No more mocked responses for the query"
errors, often causing unrelated tests to fail.

I have finally tracked the source of these errors down to the
implementation of MockLink, which was using `throw` to raise exceptions,
potentially outside the bounds of the test for which the MockLink was
originally created.

This gigantic commit does two things:

  * Make all done/done.fail-style tests use Promise resolve and reject
    functions instead (the bulk of the changes). This is important because
    we don't want stray done.fail calls causing tests that have already
    passed to fail, and resolve/reject are conveniently idempotent.

  * Make the mocking utilities (MockLink, mockSingleLink, etc.) require a
    Promise reject function, so that they can safely report errors back to
    the true original test that created the Promise.

A better testing framework would maintain an execution context for each
test that allowed exceptions to be attributed to the correct test without
explicitly passing around reject functions, but doing that right probably
requires some ECMAScript language support (or at least a transpiler to
simulate it). This is an exciting future possibility, but not something we
can tackle here.