Provide getStoreValue function to read/merge functions instead of parentObject.
Exposing the entire parent StoreObject to custom read and merge functions
was risky, even if we made it read-only, because it permitted unobserved
access to sibling fields of the parentObject, which would have prevented
effective result caching for read functions that compute their results
using sibling fields as input.

Passing in a function that takes a field name and returns the value of
that field (namely, getFieldValue) enables the same sibling field use
cases, without exposing unrelated StoreObject fields. In the future, we
can use this function to register dependencies on the fields that were
requested, which would not have been possible with just an object.