Revert "Throw if Policies#fragmentMatches receives a falsy __typename."
Fixes #5709 by reverting commit 83b58378e0e075a1ac360567b0c237386c78e9d8.

The removed test was invalid because a fragment with a type condition can
never match an object without a __typename field in Apollo Client 3.0.
The invariant was related to this test because the invariant failed
earlier, effectively hiding the true misbehavior of the test.