Support toReference(obj, true) to persist obj into store.
The options.toReference helper function is provided to custom read and
merge functions, as well as to cache.modify modifier functions, via their
details parameter (see #5909). See the tests included in this commit for
examples of all three usages.

The original purpose of toReference was to return a Reference object,
{ __ref: <ID> }, given an object with a __typename and the necessary key
fields for that type. If these expectations are not met, toReference(obj)
can return null, to indicate that the object could not be identified.

Another toReference failure mode (besides returning null) is returning a
Reference object that doesn't actually refer to anything stored in the
cache. While this case is handled gracefully while reading from the cache
(the non-existent object simply appears empty), you may have encountered
cases where you _do_ want to persist the object that was passed to
toReference into the cache, such as when a custom read function knows how
to provide a perfectly adequate default object based on options.args.

For those use cases, you can now call toReference(object, true) to merge
the fields of the object into the cache, assuming it could be identified.

The fields of the object are merged into the cache as-is, without any
transformations to handle field arguments, aliases, directives, or
variables. For those cases, cache.writeQuery or cache.writeFragment are
still the best options. Also, toReference is not exposed outside of
read/merge/modify functions, so if you're not in one of those functions
you will have to call the normal writeQuery or writeFragment methods.

Implementation-wise, since toReference may now need to call store.merge,
it was most convenient to move the definition of the toReference method
from the Policies class to the EntityStore class.