Skip writing still-fresh result objects into store.
If we're about to write a result object into the store, but we happen to
know that the exact same (===) result object would be returned if we were
to re-read the result with the same inputs right now (store, selection
set, parent object, variables), then we can skip the rest of the
processSelectionSet work for this object.

Note that this trick only works because result objects are frozen (in
development), which allows us to assume the structure of the === result
object has not changed since it was originally returned.

The executeSelectionSet.peek method is also vital here, since we need a
way to check the latest cached result without recomputing the result if
it's missing or dirty.

This optimization makes the readQuery-transform-writeQuery pattern much
cheaper when the written data contains a bunch of unmodified result
objects that we recently read from the cache.