Add extra onCompleted state change hooks test
This test was originally in the components `Query.test.tsx` file,
but for some reason it isn't working properly after the
component code was migrated into this repo. Instead of
spending more time troubleshooting it for the component code,
it's more important to verify it works with `useQuery` (which it
does).