Keep policies.rootIdsByTypename unique by ID as well as __typename.
Although the __typename of the root query, mutation, and subscription
types are almost always "Query", "Mutation", and "Subscription",
respectively, it is possible to declare different names for those root
types in your schema, which will affect queries like

  query { __typename }

This query isn't normally very remarkable or useful, except that it
happens to be the smallest legal GraphQL query.

Note that a configuration like

  new InMemoryCache({
    typePolicies: {
      RootQuery: {
        queryType: true,
      },
    },
  })

is necessary to let InMemoryCache know that "RootQuery" is the name of the
root query type, rather than "Query". This non-default behavior is not
something the cache can safely figure out without some help.

Thankfully, we have been very careful not to hard-code the default
__typename values for these root types, but #6685 demonstrates that this
generic system was not working entirely correctly.

The policies.rootTypenamesById map will always have exactly three keys,
since there are only three root IDs. However, policies.rootIdsByTypename
had a bug where more than one typename could map to the same root ID, so
you could end up with both "Query" and "RootQuery" mapping to ROOT_QUERY,
for example, which was a problem for code that needed to detect whether a
given typename corresponded to a root type or not.

We could have fixed this by getting rid of the rootIdsByTypename lookup
table, and just computing its information from rootTypenamesById whenever
we needed it, but that computation is somewhat performance-sensitive, so I
would prefer to keep the two tables in (inverse) sync with each other.

Fixes #6685.