Replace explicit externalPackages with test of non-relativity.
We don't currently have any external packages that we want to bundle into
@apollo/client, so we can get away with considering any non-relative
import to be external. The results of this test exactly match the current
contents of externalPackages, so it should be safe to remove that Set.

This should save us from making mistakes like the one I fixed in #6624,
and if we ever do want to bundle something from node_modules into the
@apollo/client bundle again, we can easily carve out an exception.