Introduce ObservableQuery#refresh as an alternative to refetch.
The difference is that refresh overwrites existing fields by not passing
existing data to field merge functions, whereas refetch passes existing
data to merge functions, thereby combining it with the refetch result.

If we continue down this path, we can say "the solution to #7491 is to
switch from using refetch to using refresh," but that involves conscious
effort by the developer, and would mean maintaining two very similar
methods of the ObservableQuery class (refetch and refresh).