Make optimistic CacheGroup inherit from non-optimistic CacheGroup.
This change means the cached results of optimistic reads can be
invalidated by non-optimistic writes, which makes sense because the
non-optimistic writes potentially affect data that was previously
inherited by optimistic layers and consumed by optimistic reads.

I'm not convinced this is absolutely necessary, but it's generally safe to
err on the side of over-invalidation.