Improve options API of client.refetchQueries method.
This is obviously a breaking change for client.refetchQueries, but that
method was introduced in PR #7431, targeting the release-3.4 branch.

Since Apollo Client v3.4 is still in beta, we still have room to rework
the signature of the client.refetchQueries method (introduced in #7431 and
released in @apollo/client@3.4.0-beta.3), not only adding significant new
functionality like options.updateCache and options.onQueryUpdated, but
also leaving room to add functionality more easily in the future, without
breaking backwards compatibility, since client.refetchQueries takes named
options and returns an object of named results, so adding new options or
returning new results never needs to be a breaking change.